Time: 1 ms (84.3%), Space: 41 MB (53.98%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -8,7 +8,7 @@ class Solution {
 
     while(start<=end){
         int mid = start + (end-start) /2;
-        int square = (long) mid*mid;
+        long square = (long) mid*mid;
         if(square == x){
             return mid;
         }
